refactor: Use Casting float to int
pandas로 추가 타입 지정을 했지만
Java에서 csv를 읽어올 때 object를 float으로 읽어와서
int 로 캐스팅해서 set함

+++ src/main/java/com/springboothospitalapi/domain/Hospital.java
@@ -24,8 +24,8 @@ public class Hospital {
     private String roadNameAddress;
     private String hospitalName;
     private String businessTypeName;
-    private float healthcareProviderCount;
-    private float patientRoomCount;
-    private float totalNumberOfBeds;
+    private int healthcareProviderCount;
+    private int patientRoomCount;
+    private int totalNumberOfBeds;
     private float totalAreaSize;
 }
